+++ lol_stocks_web_server/src/endpoints/mod.rs
@@ -2,5 +2,5 @@ pub mod register_games;
 pub mod padlock;
 pub mod register_teams;
 pub mod teams;
-pub mod take_team_snapshot;
+pub mod take_snapshot;
 pub mod make_admin;
