+++ lol_stocks_web_server/src/endpoints/take_snapshot.rs
